+++ test/e2e/e2e_test.go
@@ -54,6 +54,11 @@ func TestFinalityContractTestSuite(t *testing.T) {
 	suite.Run(t, new(FinalityContractTestSuite))
 }
 
+// TestIbcCallbackBsnAddRewards tests BSN fee collection via IBC callbacks end-to-end
+func TestIbcCallbackBsnAddRewards(t *testing.T) {
+	suite.Run(t, new(IbcCallbackBsnAddRewards))
+}
+
 // TestBtcRewardsDistributionBsnRollup tests the bsn rewards for rollups
 func TestBtcRewardsDistributionBsnRollup(t *testing.T) {
 	suite.Run(t, new(BtcRewardsDistributionBsnRollup))