+++ x/btcstkconsumer/keeper/params.go
@@ -3,31 +3,20 @@ package keeper
 import (
 	"context"
 
-	"github.com/cosmos/cosmos-sdk/runtime"
-
 	"github.com/babylonlabs-io/babylon/v3/x/btcstkconsumer/types"
 )
 
 // GetParams get all parameters as types.Params
 func (k Keeper) GetParams(ctx context.Context) (params types.Params) {
-	store := runtime.KVStoreAdapter(k.storeService.OpenKVStore(ctx))
-	bz := store.Get(types.ParamsKey)
-	if bz == nil {
+	result, err := k.ParamsCollection.Get(ctx)
+	if err != nil {
+		// Return default params if not found
 		return params
 	}
-
-	k.cdc.MustUnmarshal(bz, &params)
-	return params
+	return result
 }
 
 // SetParams set the params
 func (k Keeper) SetParams(ctx context.Context, params types.Params) error {
-	store := runtime.KVStoreAdapter(k.storeService.OpenKVStore(ctx))
-	bz, err := k.cdc.Marshal(&params)
-	if err != nil {
-		return err
-	}
-	store.Set(types.ParamsKey, bz)
-
-	return nil
+	return k.ParamsCollection.Set(ctx, params)
 }