+++ CHANGELOG.md
@@ -39,6 +39,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Improvements
 
+- [#1406](https://github.com/babylonlabs-io/babylon/pull/1406) Use `collections` for KV store in `x/btcstkconsumer`.
+- [#1405](https://github.com/babylonlabs-io/babylon/pull/1405) Use `collections` for `zoneconcierge` KV store.
 - [#1387](https://github.com/babylonlabs-io/babylon/pull/1387) Refactor the usage of `ChainInfo` and canonical chain indexing.
 - [#1331](https://github.com/babylonlabs-io/babylon/pull/1331/) Improve BSN Header synchronization.
 - [#1304](https://github.com/babylonlabs-io/babylon/pull/304) Update rollup finality e2e tests
@@ -118,6 +120,8 @@ command to modify `max-finality-providers` parameter.
 - [#1381](https://github.com/babylonlabs-io/babylon/pull/1381) Add multi-staking allow-list logic
 - [#1388](https://github.com/babylonlabs-io/babylon/pull/1388) Upgrade: state migration for v3
 - [#1392](https://github.com/babylonlabs-io/babylon/pull/1392) Add e2e test for BSN rollup rewards.
+- [#1303](https://github.com/babylonlabs-io/babylon/pull/1303) Add IBC callback middleware to `x/btcstaking` module to call add bsn rewards.
+- [#1404](https://github.com/babylonlabs-io/babylon/pull/1404) Refactor power dist event processing.
 
 ### State Machine Breaking
 
@@ -143,6 +147,8 @@ cost per multi-staked finality provider
 FP data
 - [#1398](https://github.com/babylonlabs-io/babylon/pull/1398) Add custom tokenfactory
 bindings (thanks @benluelo for contribution)
+- [#1416](https://github.com/babylonlabs-io/babylon/pull/1416) Releax requirements
+for stake expansion funding output
 
 ### Bug fixes
 
@@ -168,6 +174,8 @@ already expanded delegations
 for multi staked fps
 - [#1384](https://github.com/babylonlabs-io/babylon/pull/1384) fix: Cascaded slashing
 adjustment
+- [#1418](https://github.com/babylonlabs-io/babylon/pull/1418) fix: add stake expansion
+covenant signature to the events
 
 ## v2.2.0